services/ocr_service/pyproject.toml
[project]
name = "ocr-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.10.1",
    "fastapi[standard]>=0.116.1",
    "pillow>=11.3.0",
    "python-multipart>=0.0.20",
    "sentence-transformers>=5.1.0",
    "tokenizers>=0.22.0",
    "torch>=2.8.0",
    "transformers>=4.56.1",
    "uvicorn>=0.35.0",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=5.1.0",
    "sqlalchemy>=2.0.43",
    "torchaudio>=2.8.0",
    "torchvision>=0.23.0",
    "llama-cpp-python>=0.3.16",
    "huggingface-hub>=0.34.4",
    "pymupdf>=1.26.4",
]
